made cannon mass to int

+++ src/main/java/at/pavlov/cannons/cannon/DesignStorage.java
@@ -221,7 +221,7 @@ public class DesignStorage
 		cannonDesign.setHasRecoil(cannonDesignConfig.getBoolean("realisticBehaviour.hasRecoil", false));
 		cannonDesign.setFrontloader(cannonDesignConfig.getBoolean("realisticBehaviour.isFrontloader", false));
 		cannonDesign.setRotatable(cannonDesignConfig.getBoolean("realisticBehaviour.isRotatable", false));
-        cannonDesign.setMassOfCannon(cannonDesignConfig.getDouble("realisticBehaviour.massOfCannon", 1000.0));
+        cannonDesign.setMassOfCannon(cannonDesignConfig.getInt("realisticBehaviour.massOfCannon", 1000.0));
 
 		// permissions
 		cannonDesign.setPermissionBuild(cannonDesignConfig.getString("permissions.build", "cannon.player.build"));
